fix-kmitl_demo-init_lang
