+++ kmitl_demo/__manifest__.py
@@ -18,8 +18,10 @@
         "data/company.xml",
         "data/hr.department.csv",
         "data/account.fiscal.year.csv",
-        "data/ir_config_parameter.xml"
+        "data/ir_config_parameter.xml",
     ],
+    "post_init_hook": "post_init",
+    "uninstall_hook": "uninstall_hook",
     "application": False,
     "installable": True,
     "auto_install": False,
