+++ kmitl_demo/hooks.py
@@ -1,12 +1,38 @@
-from odoo import api, SUPERUSER_ID
+from odoo import SUPERUSER_ID, api
+from odoo.fields import Command
 
-def post_init_hook(cr, registry):
+
+def post_init(cr, registry):
+    # This is a placeholder for the post-init method.
     env = api.Environment(cr, SUPERUSER_ID, {})
-    # โหลดภาษาไทยถ้ายังไม่ถูกโหลด
-    if 'th_TH' not in env['res.lang'].search([('code', '=', 'th_TH')]).mapped('code'):
-        env['res.lang'].load_lang('th_TH')
-    # ตั้งค่าภาษาให้กับผู้ใช้ทุกคน
-    users = env['res.users'].search([])
-    users.write({'lang': 'th_TH'})
-    # ตรวจสอบให้แน่ใจว่า th_TH ถูกเปิดใช้งาน
-    env['res.lang']._activate_lang('th_TH')
+
+    # Install the Thai language pack
+    th = (
+        env["res.lang"].with_context(active_test=False).search([("code", "=", "th_TH")])
+    )
+    ll = env["base.language.install"].create(
+        {"lang_ids": [Command.clear(), Command.link(th.id)]}
+    )
+    ll.lang_install()
+
+    # Set the default language for the system to Thai
+    default_lang = env["ir.default"].browse([1])
+    default_lang.json_value = '"th_TH"'
+
+    # Set the default language for all users to Thai
+    users = env["res.users"].search([])
+    users.lang = "th_TH"
+
+
+def uninstall_hook(cr, registry):
+    # This is a placeholder for the uninstall method.
+    env = api.Environment(cr, SUPERUSER_ID, {})
+
+    default_lang = env["ir.default"].browse([1])
+    default_lang.json_value = '"en_US"'
+
+    users = env["res.users"].search([])
+    users.lang = "en_US"
+
+    rl = env["res.lang"].search([("code", "=", "th_TH")])
+    rl.active = False
